Move CodeNode into nodes/data and register it in the catalog
CodeNode is the customisation port for user-authored data-transform
logic, so it belongs alongside the other data nodes rather than at
the nodes package root. Also register it in the node registry (category
"data") so it appears in `orcheo node list`, matching the NoOpTaskNode
discovery pattern.
